+++ client/src/Pages/BuildProject/BuildProject.tsx
@@ -2,10 +2,11 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { FaCamera } from 'react-icons/fa';
 import { sampleImages } from '../../Components/BuildProject/sampleImages';
-import { FieldData, PosData } from '../../Components/Common/OptionData';
+import { FieldData } from '../../Components/Common/OptionData';
 import Button from '../../Components/Common/Button';
 import SelectBox from '../../Components/Common/SelectBox';
 import InputBox from '../../Components/Common/InputBox';
+import RecruitSelect from '../../Components/BuildProject/RecruitSelect';
 import DatePicker from 'react-datepicker';
 import { ko } from 'date-fns/esm/locale';
 import 'react-datepicker/dist/react-datepicker.css';
@@ -136,49 +137,6 @@ const SDatePicker = styled(DatePicker)`
   font-size: 12px;
 `;
 
-const RecruitWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const RecruitSelect = styled.select`
-  width: 200px;
-  height: 42px;
-  border-radius: 4px;
-  border: 1px solid ${(props) => props.theme.palette.lightGray};
-  padding-left: 20px;
-  margin-right: 15px;
-  font-size: 12px;
-`;
-
-const RecruitOption = styled.option`
-  text-align: center;
-  margin: 1.2rem;
-  color: ${(props) => props.theme.palette.gray};
-  padding: 0px 2px 1px;
-`;
-
-const CountWrapper = styled.div`
-  justify-content: center;
-  margin: 0 1.5rem;
-  font-size: 20px;
-`;
-
-const PosCount = styled.span`
-  color: ${(props) => props.theme.palette.orange};
-  margin: 0 1.5rem;
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const PosCountBtn = styled.button`
-  background: none;
-  border: none;
-  user-select: none;
-  font-size: 20px;
-`;
-
 const TwoBtnWrapper = styled.div`
   & button {
     margin-left: 0;
@@ -218,6 +176,18 @@ function BuildProject() {
 
   const handleClickRadioButton = (value: string) => setField(value);
 
+  const handleAddPosClick = () => {
+    setPositions([...positions, { pos: 'none', count: 1 }]);
+  };
+
+  const handleDeletePosClick = () => {
+    if (positions.length !== 1) {
+      const newPositons = [...positions];
+      newPositons.splice(-1, 1);
+      setPositions(newPositons);
+    }
+  };
+
   const handleReferenceChange = (
     event: React.ChangeEvent<HTMLInputElement>,
     index: number
@@ -239,45 +209,6 @@ function BuildProject() {
       const newReferences = [...referencesUrl];
       newReferences.splice(-1, 1);
       setReferencesUrl(newReferences);
-    }
-  };
-
-  const handlePosSelectChange = (
-    event: React.ChangeEvent<HTMLSelectElement>,
-    index: number
-  ) => {
-    const { value } = event.target;
-    const newPositions = [...positions];
-    newPositions[index].pos = value;
-    setPositions(newPositions);
-  };
-
-  const handlePosCountClick = (
-    event: React.MouseEvent<HTMLButtonElement>,
-    index: number
-  ) => {
-    event.preventDefault();
-    const { innerText: type } = event.target as HTMLButtonElement;
-    const newPositions = [...positions];
-    if (type === '+') {
-      newPositions[index].count = newPositions[index].count + 1;
-    } else {
-      if (newPositions[index].count !== 1) {
-        newPositions[index].count = newPositions[index].count - 1;
-      }
-    }
-    setPositions(newPositions);
-  };
-
-  const addPosClickHandler = () => {
-    setPositions([...positions, { pos: 'none', count: 1 }]);
-  };
-
-  const removePosClickHandler = () => {
-    if (positions.length !== 1) {
-      const newPositons = [...positions];
-      newPositons.splice(-1, 1);
-      setPositions(newPositons);
     }
   };
 
@@ -400,43 +331,28 @@ function BuildProject() {
             * 나중에 변경/추가가 가능합니다. 3~4명을 추천합니다.
           </SectionInfo>
           {positions.map((pos, index) => (
-            <RecruitWrapper key={index}>
-              <RecruitSelect
-                value={pos.pos}
-                onChange={(e) => handlePosSelectChange(e, index)}
-              >
-                {PosData.map((item, index) => (
-                  <RecruitOption key={index} value={item.value}>
-                    {item.label}
-                  </RecruitOption>
-                ))}
-              </RecruitSelect>
-              <CountWrapper>
-                <PosCountBtn onClick={(e) => handlePosCountClick(e, index)}>
-                  -
-                </PosCountBtn>
-                <PosCount>{pos.count}</PosCount>
-                <PosCountBtn onClick={(e) => handlePosCountClick(e, index)}>
-                  +
-                </PosCountBtn>
-              </CountWrapper>
-            </RecruitWrapper>
+            <RecruitSelect
+              key={index}
+              positions={positions}
+              value={pos}
+              submitValue={setPositions}
+              index={index}
+            />
           ))}
-
           <TwoBtnWrapper>
             <Button
               ButtonColor="darkblue"
               ButtonMode="active"
               ButtonSize="small"
               ButtonName="삭제"
-              onClick={removePosClickHandler}
+              onClick={handleDeletePosClick}
             />
             <Button
               ButtonColor="white"
               ButtonMode="active"
               ButtonSize="small"
               ButtonName="추가"
-              onClick={addPosClickHandler}
+              onClick={handleAddPosClick}
             />
           </TwoBtnWrapper>
         </Section>
